Moved src files into new vda5050_connector
